Moved midblock adapter to its own module
