+++ distvae/modules/adapters/vae/decoder_adapters.py
@@ -7,12 +7,11 @@ from torch.distributed import ProcessGroup
 from torch.profiler import profile, ProfilerActivity
 
 from distvae.models.vae import PatchDecoder
-from distvae.modules.adapters.layers.attn_adapters import WanAttentionBlockAdapter
 from distvae.modules.adapters.layers.conv_adapters import Conv2dAdapter, WanCausalConv3dAdapter
 from distvae.modules.adapters.layers.norm_adapters import GroupNormAdapter
-from distvae.modules.adapters.resnet_adapters import WanResidualBlockAdapter
 from distvae.modules.adapters.unets.unet_2d_blocks_adapters import UpDecoderBlock2DAdapter
 from distvae.modules.adapters.upsampling_adapters import WanResidualUpBlockAdapter, WanUpBlockAdapter
+from distvae.modules.adapters.midblock_adapters import WanMidBlockAdapter
 from distvae.modules.patch_utils import Patchify, DePatchify
 from distvae.utils import DistributedEnv
 from diffusers.models.autoencoders.vae import Decoder
@@ -93,26 +92,6 @@ class DecoderAdapter(nn.Module):
         if rank == 0:
             print(f"Patch vae: [elapsed_time: {elapsed_time:.2f} sec, peak_memory: {peak_memory/1e9} GB]")
         return output
-
-
-class WanMidBlockAdapter(nn.Module):
-    def __init__(self, wan_mid_block: WanMidBlock, conv_block_size = 0, patch_dim: int = -2):
-        super().__init__()
-
-        assert isinstance(wan_mid_block, WanMidBlock), "WanMidBlockAdapter does not support mid block except WanMidBlock"
-        self.mid_block = wan_mid_block
-        self.mid_block.resnets = nn.ModuleList([
-            WanResidualBlockAdapter(
-                resnet, conv_block_size=conv_block_size, patch_dim=patch_dim
-            ) for resnet in wan_mid_block.resnets
-        ])
-        self.mid_block.attentions = nn.ModuleList([
-            WanAttentionBlockAdapter(attn, patch_dim=patch_dim)
-            for attn in wan_mid_block.attentions
-        ])
-
-    def forward(self, x, feat_cache=None, feat_idx=[0]):
-        return self.mid_block(x, feat_cache=feat_cache, feat_idx=feat_idx)  
 
 
 class WanDecoderAdapter(nn.Module):
